Create sudokuToSAT.py
The first program of the Sudoku solver.
It reads one n by n puzzle. It creates a list of the characters used in the puzzle assigning them ids.
Using the ids it saves the locations and ids of the known parts of the puzzle.
The program will write the minimal miniSAT solver file.
The program will also write a meta file that preserves the id's and characters for printing the solved puzzle. If the puzzle is 9 by 9 and uses characters 1-9 the meta file will only say 'Skip Meta'. 

Currently the miniSAT solver file is not written.